Update intro.rst
"It is slow to send each IO in test scripts, so pynvme provides a mocked agent to send IOes in separated processes." 

actually I don't fully understand why agents in separated processes makes it no longer slow.  Is it "to send it is slow", or "the IO in this way is slow"?  Or did you mean to claim with one single process, the throughput would be capped low?